+++ lib_gdapi.py
@@ -6,45 +6,13 @@ from google_auth_oauthlib.flow import InstalledAppFlow
 from google.auth.transport.requests import Request
 from googleapiclient.http import MediaFileUpload, MediaIoBaseDownload
 import io
+import re
 # If modifying these scopes, delete the file token.pickle.
 
 SCOPES = ['https://www.googleapis.com/auth/drive']
 #mimetype dict
 PickleFile=r'D:\Python\All_Practice\GoogleAPI\apitesla_token.pickle'
 mimetype_dict={'xls': 'application/vnd.ms-excel', 'xlsx': 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet', 'xml': 'text/xml', 'ods': 'application/vnd.oasis.opendocument.spreadsheet', 'csv': 'text/csv', 'tmpl': 'text/plain', 'pdf': 'application/pdf', 'php': 'application/x-httpd-php', 'jpg': 'image/jpeg', 'png': 'image/png', 'gif': 'image/gif', 'bmp': 'image/bmp', 'txt': 'text/plain', 'doc': 'application/msword', 'js': 'text/js', 'swf': 'application/x-shockwave-flash', 'mp3': 'audio/mpeg', 'zip': 'application/zip', 'rar': 'application/rar', 'tar': 'application/tar', 'arj': 'application/arj', 'cab': 'application/cab', 'html': 'text/html', 'htm': 'text/html', 'default': 'application/octet-stream', 'folder': 'application/vnd.google-apps.folder', '': 'application/vnd.google-apps.video', 'Google Docs': 'application/vnd.google-apps.document', '3rd party shortcut': 'application/vnd.google-apps.drive-sdk', 'Google Drawing': 'application/vnd.google-apps.drawing', 'Google Drive file': 'application/vnd.google-apps.file', 'Google Drive folder': 'application/vnd.google-apps.folder', 'Google Forms': 'application/vnd.google-apps.form', 'Google Fusion Tables': 'application/vnd.google-apps.fusiontable', 'Google Slides': 'application/vnd.google-apps.presentation', 'Google Apps Scripts': 'application/vnd.google-apps.script', 'Shortcut': 'application/vnd.google-apps.shortcut', 'Google Sites': 'application/vnd.google-apps.site', 'Google Sheets': 'application/vnd.google-apps.spreadsheet'}
-
-def main():
-    
-    """Shows basic usage of the Drive v3 API.
-    Prints the names and ids of the first 10 files the user has access to.
-    """
-    creds = None
-    # The file token.pickle stores the user's access and refresh tokens, and is
-    # created automatically when the authorization flow completes for the first
-    # time.
-    if os.path.exists(PickleFile):
-        with open(PickleFile, 'rb') as token:
-            creds = pickle.load(token)
-    # If there are no (valid) credentials available, let the user log in.
-    if not creds or not creds.valid:
-        if creds and creds.expired and creds.refresh_token:
-            creds.refresh(Request())
-        else:
-            flow = InstalledAppFlow.from_client_secrets_file(
-                'credentials.json', SCOPES)
-            creds = flow.run_local_server(port=8012)
-        # Save the credentials for the next run
-        with open(PickleFile, 'wb') as token:
-            pickle.dump(creds, token)
-
-    service = build('drive', 'v3', credentials=creds)
-    
-
-    # Call the Drive v3 API#
-   
-    
-    return service
-service=main()
 def get_minitype_txt_todict(f=r'mimetype.txt'):# get mimetype dict from  txt
     dict_mime={}
     with open(f,'r') as ff:
@@ -61,17 +29,7 @@ def get_minitype_txt_todict(f=r'mimetype.txt'):# get mimetype dict from  txt
                         bb=bb.replace("'",'').replace(",",'').strip()
                         dict_mime[bb]=aa
     return dict_mime
-def get_dict(dict_form):
-    
-    newdict={}
-    # dict_form={'a':'b'} , this fun will return a dict
-    split1=dict_form.split(',')
-    for i in split1:
-            a,b=i.split(':')
-            a=a.replace('{','').replace("'",'')
-            b=b.replace('{','').replace("'",'')
-            newdict[a]=b
-    return newdict
+
     
 
 def find_folder():
@@ -88,7 +46,7 @@ def find_file(filename):
       
       
       items = results.get('files', [])
-      
+     
       
       if not items:return 0
       
@@ -141,8 +99,8 @@ def download(file,dst):
         with io.open(filepath,'wb') as f:
             fh.seek(0)
             f.write(fh.read())
-def change_permissions():# 將某ID的檔案權限更改
-    service.permissions().create(fileId='1wgobAC45IMM-TUE58fO1P8D8_FURMXyLjRXzPaKI_pQ',
+def change_permissions(file_id):# 將某ID的檔案權限更改
+    service.permissions().create(fileId=file_id,
                                          body= {
                                         'role': 'writer',
                                         'type': 'anyone',
@@ -154,14 +112,11 @@ def create_newsheet(filename):
     'name': filename,
     'mimeType': 'application/vnd.google-apps.spreadsheet'
     }
-   
+    
     file = service.files().create(body=file_metadata).execute()
-def upload(filename,file_type):
-    file_metadata = {'name': filename}
-    media = MediaFileUpload(filename, mimetype=file_type)
-    file = service.files().create(body=file_metadata,
-                                        media_body=media,
-                                        fields='id').execute()
+    
+    return file['id']
+
 def find_folder_id(foldername):
       results = service.files().list(q="mimeType='application/vnd.google-apps.folder'",
                                           spaces='drive',
@@ -209,7 +164,7 @@ def upload_file(filepath,dstpath):# 上傳檔案到特地資料夾
         if Extension in mimetype_dict:
             file_type=mimetype_dict[Extension]
     else:
-        print("Mimetype not found , try to upload to this folder")
+        print("Mimetype not found , try to upload this folder")
         foldername=os.path.basename(dstpath)
         filename=os.path.basename(filepath)
         Extension=os.path.splitext(filename)[-1].replace('.','')
@@ -244,44 +199,83 @@ def upload_file(filepath,dstpath):# 上傳檔案到特地資料夾
     file = service.files().create(body=file_metadata,
                                         media_body=media,
                                         fields='id').execute()
-    print("Finish !")
+    print("Upload Finish !")
 def delete_file(filename):
   file_id=find_file(filename)
+  if not file_id:
+      print("Not found file to delete")
+      return
   """Permanently delete a file, skipping the trash
 
   Args:
     service: Drive API service instance.
     file_id: ID of the file to delete.
   """
+  file_=service.files().get(fileId=file_id).execute()
+  print("Delete: ",file_['name'])
   try:
         service.files().delete(fileId=file_id).execute()
   except :
       print("Delete Error")
+def get_():
+    # fields = get what you want
+    # file resource url=https://developers.google.com/drive/api/v3/reference/files
+    results = service.files().get(fileId=fileid,fields='webViewLink').execute()
+def main():
+    """Shows basic usage of the Drive v3 API.
+    Prints the names and ids of the first 10 files the user has access to.
+    """
+    creds = None
+    # The file token.pickle stores the user's access and refresh tokens, and is
+    # created automatically when the authorization flow completes for the first
+    # time.
+    if os.path.exists(PickleFile):
+        with open(PickleFile, 'rb') as token:
+            creds = pickle.load(token)
+    # If there are no (valid) credentials available, let the user log in.
+    if not creds or not creds.valid:
+        if creds and creds.expired and creds.refresh_token:
+            creds.refresh(Request())
+        else:
+            flow = InstalledAppFlow.from_client_secrets_file(
+                'credentials.json', SCOPES)
+            creds = flow.run_local_server(port=8012)
+        # Save the credentials for the next run
+        with open(PickleFile, 'wb') as token:
+            pickle.dump(creds, token)
+
+    service = build('drive', 'v3', credentials=creds)
+    service_sheet = build('sheets', 'v4', credentials=creds)
+         
+ 
+    # Call the Drive v3 API#
+   
     
+    return service,service_sheet
 
-
+def write(sheetid,workRange,content):
+    sheet = service_sheet.spreadsheets()
+    writeData = {}
+    
+ 
+    writeData['values']= [[content]]
+    writeData['majorDimension']="ROWS"
+    writeData['range']= workRange
+    
+   
+    
+    request = service_sheet.spreadsheets().values().update(spreadsheetId=sheetid, range=workRange,valueInputOption='RAW',body=writeData )
+    response = request.execute()
+service,service_sheet = main()
 if __name__ == '__main__':
-     service = main()
-
-   # = Upload =
-   
-    # src=r'D:\Python\All_Practice\GoogleAPI\test/test.txt'
-    # dst='test.txt'
-    
-    # upload_file(src,dst)# src /to dst (若沒給會自動偵測)//資料夾ID
-    
-   # = download =
-   
-     #file_='....txt'
-     #dst=r'D:\Python\All_Practice\GoogleAPI\test'
-     # download(file_,dst)
+     service,service_sheet = main()
      
-   # = delete =
-     #delete_file(filename)
-   
    # = create sheet =
    
-    # create_newsheet("HI信登")
+    # sheetid=create_newsheet("HI")
+     sheetid='18i20TCq8ujAELdMTCmZpazXIVBwIX02_6SeAHYl7pMs'
+     
+     write(sheetid,'H7','123HI')
